+++ Engine/ac/dynobj/cc_dynamicarray.cpp
@@ -63,7 +63,7 @@ void CCDynamicArray::Unserialize(int index, const char *serializedData, int data
     ccRegisterUnserializedObject(index, &newArray[8], this);
 }
 
-long CCDynamicArray::Create(int numElements, int elementSize, bool isManagedType)
+int32_t CCDynamicArray::Create(int numElements, int elementSize, bool isManagedType)
 {
     char *newArray = new char[numElements * elementSize + 8];
     memset(newArray, 0, numElements * elementSize + 8);
@@ -76,52 +76,52 @@ long CCDynamicArray::Create(int numElements, int elementSize, bool isManagedType
 }
 
 
-void CCDynamicArray::Read(const char *address, int offset, void *dest, int size)
+void CCDynamicArray::Read(const char *address, intptr_t offset, void *dest, int size)
 {
     memcpy(dest, address + offset, size);
 }
 
-uint8_t CCDynamicArray::ReadInt8(const char *address, long offset)
+uint8_t CCDynamicArray::ReadInt8(const char *address, intptr_t offset)
 {
     return *(uint8_t*)(address + offset);
 }
 
-int16_t CCDynamicArray::ReadInt16(const char *address, long offset)
+int16_t CCDynamicArray::ReadInt16(const char *address, intptr_t offset)
 {
     return *(int16_t*)(address + offset);
 }
 
-int32_t CCDynamicArray::ReadInt32(const char *address, long offset)
+int32_t CCDynamicArray::ReadInt32(const char *address, intptr_t offset)
 {
     return *(int32_t*)(address + offset);
 }
 
-float CCDynamicArray::ReadFloat(const char *address, long offset)
+float CCDynamicArray::ReadFloat(const char *address, intptr_t offset)
 {
     return *(float*)(address + offset);
 }
 
-void CCDynamicArray::Write(const char *address, int offset, void *src, int size)
+void CCDynamicArray::Write(const char *address, intptr_t offset, void *src, int size)
 {
     memcpy((void*)(address + offset), src, size);
 }
 
-void CCDynamicArray::WriteInt8(const char *address, long offset, uint8_t val)
+void CCDynamicArray::WriteInt8(const char *address, intptr_t offset, uint8_t val)
 {
     *(uint8_t*)(address + offset) = val;
 }
 
-void CCDynamicArray::WriteInt16(const char *address, long offset, int16_t val)
+void CCDynamicArray::WriteInt16(const char *address, intptr_t offset, int16_t val)
 {
     *(int16_t*)(address + offset) = val;
 }
 
-void CCDynamicArray::WriteInt32(const char *address, long offset, int32_t val)
+void CCDynamicArray::WriteInt32(const char *address, intptr_t offset, int32_t val)
 {
     *(int32_t*)(address + offset) = val;
 }
 
-void CCDynamicArray::WriteFloat(const char *address, long offset, float val)
+void CCDynamicArray::WriteFloat(const char *address, intptr_t offset, float val)
 {
     *(float*)(address + offset) = val;
 }
